Now you can write a journal entry and save it.
It will be saved in sharedpreferences.

When you save it, you can see it on a list on the main screen

You can delete items now by long clicking on a entry,
it will be saved to sharedpreferences too.

The last thing I need to implement is sorting the entries by timestamp